// apps/desktop/src/lib/types/Participant.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Participant = { identity: string, name: string, 
/**
 * We receive a voice track from them (otherwise they are in the room but silent to us).
 */
hasVoice: boolean, 
/**
 * They muted their own microphone.
 */
muted: boolean, };